refactor file does not exist bug fix

+++ server/commands.rb
@@ -101,23 +101,23 @@ class Commands
       splitter = '\\'
     end
     client.puts "put #{file.split(splitter).last} #{format}"
-    begin
-      if format == 'binary'
-        file = File.open(file, 'rb')
-      elsif format == 'text'
-        file = File.open(file, 'r')
-      else
-        puts 'invalid format'
-      end
-      content = file.gets
-      file.close
-      client.puts content
+    unless File.exist? file
+      puts 'File does not exist!'
       client.puts 'done'
-      puts 'upload finished'
-    rescue Errno::ENOENT
-      puts 'That file does not exist'
-      client.puts 'done'
+      return
     end
+    if format == 'binary'
+      file = File.open(file, 'rb')
+    elsif format == 'text'
+      file = File.open(file, 'r')
+    else
+      puts 'invalid format'
+    end
+    content = file.gets
+    file.close
+    client.puts content
+    client.puts 'done'
+    puts 'upload finished'
   end
 
   def self.scan(client)
